+++ packages/react/src/components/LeadspaceWithSearch/__stories__/LeadspaceWithSearch.stories.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2016, 2021
+ * Copyright IBM Corp. 2016, 2022
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -11,53 +11,40 @@ import React from 'react';
 import readme from '../README.stories.mdx';
 import { text } from '@storybook/addon-knobs';
 
+const props = () => ({
+  heading: text('Heading (heading):', 'Find a product'),
+  copy: text('Copy (copy):', 'Duis aute irure dolor in reprehen deritirure'),
+  searchProps: {
+    placeholder: {
+      mobile: text(
+        'Mobile placeholder (searchProps.placeHolder.mobile):',
+        'Search keywords'
+      ),
+      desktop: text(
+        'Desktop placeholder (searchProps.placeHolder.desktop)',
+        'Search known issues, documentation and support forums'
+      ),
+    },
+    labelText: 'Search keywords',
+    onChange: () => {},
+    onKeyDown: () => {},
+  },
+});
+
 export default !DDS_LEADSPACE_WITH_SEARCH
   ? undefined
   : {
-      title: 'Components|Lead space with search',
+      title: 'Components/Lead space with search',
       parameters: {
         ...readme.parameters,
-        knobs: {
-          LeadspaceWithSearch: ({ groupId }) => ({
-            heading: text('Heading (heading):', 'Find a product', groupId),
-            copy: text(
-              'Copy (copy):',
-              'Duis aute irure dolor in reprehen deritirure',
-              groupId
-            ),
-            searchProps: {
-              placeholder: {
-                mobile: text(
-                  'Mobile placeholder (searchProps.placeHolder.mobile):',
-                  'Search keywords',
-                  groupId
-                ),
-                desktop: text(
-                  'Desktop placeholder (searchProps.placeHolder.desktop)',
-                  'Search known issues, documentation and support forums',
-                  groupId
-                ),
-              },
-              labelText: 'Search keywords',
-              onChange: () => {},
-              onKeyDown: () => {},
-            },
-          }),
+        percy: {
+          skip: true,
         },
       },
     };
 
 export const Default = !DDS_LEADSPACE_WITH_SEARCH
   ? undefined
-  : ({ parameters }) => {
-      const { heading, copy, searchProps } =
-        parameters?.props?.LeadspaceWithSearch ?? {};
-
-      return (
-        <LeadspaceWithSearch
-          heading={heading}
-          copy={copy}
-          searchProps={searchProps}
-        />
-      );
+  : () => {
+      return <LeadspaceWithSearch {...props()} />;
     };